requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Google AI and ADK
google-adk>=1.0.0
google-generativeai>=0.3.0


# Data Processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Utilities
python-multipart>=0.0.6  # for file uploads
aiofiles>=23.0.0  # for async file operations

# Development (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Security
python-jose[cryptography]>=3.3.0  # if using JWT tokens
passlib[bcrypt]>=1.7.4  # if using password hashing

# HTTP Client (for external APIs)
httpx>=0.25.0
requests>=2.31.0

# Image Processing (for receipt processing)
Pillow>=10.0.0

# JSON handling
orjson>=3.9.0  # faster JSON processing